refactor and updated readme

+++ install.py
@@ -8,15 +8,17 @@ def install():
     if not maya_app_dir or not os.path.isdir(maya_app_dir):
         return
 
-    directories = [
-        '2022',
-        'scripts']
-
     current_directory = os.path.dirname(os.path.abspath(__file__))
 
-    for directory in directories:
-        source = os.path.join(current_directory, directory)
-        target = os.path.join(maya_app_dir, directory)
+    for file in os.listdir(current_directory):
+        if file.startswith('.'):
+            continue
+
+        if os.path.isfile(os.path.join(current_directory, file)):
+            continue
+
+        source = os.path.join(current_directory, file)
+        target = os.path.join(maya_app_dir, file)
 
         shutil.copytree(source, target, dirs_exist_ok=True)
